Periodically refresh AMQP and inotify handles
We also had a stray `forever` which didn't do the intended dump of all
files every delayedRequeueHours. As well as fixing the bug, instead of
dumping all files, we'll now refresh the AMQP connection and inotify
handle.

We should also use addConnectionClosedHandler at some point, instead of
relying on the delayedRequeueHours configuration property.